pyproject.toml
[tool.poetry]
name = "poetryenv"
version = "0.1.0"
description = "Virtual Environment and dependency holder"
authors = ["Jaefer-Kemal <jaeferkemal@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
